// src/components/echart/types.d.ts
interface GridOption extends ComponentOption, BoxLayoutOptionMixin, ShadowOptionMixin {
  mainType?: 'grid';
  show?: boolean;
  containLabel?: boolean;
  backgroundColor?: ZRColor;
  borderWidth?: number;
  borderColor?: ZRColor;
  tooltip?: any;
}

declare type CartesianAxisPosition = 'top' | 'bottom' | 'left' | 'right';
declare type CartesianAxisOption = AxisBaseOption & {
  gridIndex?: number;
  gridId?: string;
  position?: CartesianAxisPosition;
  offset?: number;
  categorySortInfo?: OrdinalSortInfo;
};
declare type XAXisOption = CartesianAxisOption & {
  mainType?: 'xAxis';
};
declare type YAXisOption = CartesianAxisOption & {
  mainType?: 'yAxis';
};
// ...
interface LegendOption extends ComponentOption, LegendStyleOption, BoxLayoutOptionMixin, BorderOptionMixin {
  mainType?: 'legend';
  show?: boolean;
  orient?: LayoutOrient;
  align?: 'auto' | 'left' | 'right';
  backgroundColor?: ColorString;
  /**
   * Border radius of background rect
   * @default 0
   */
  borderRadius?: number | number[];
  /**
   * Padding between legend item and border.
   * Support to be a single number or an array.
   * @default 5
   */
  padding?: number | number[];
  /**
   * Gap between each legend item.
   * @default 10
   */
  itemGap?: number;
  /**
   * Width of legend symbol
   */
  itemWidth?: number;
  /**
   * Height of legend symbol
   */
  itemHeight?: number;
  selectedMode?: boolean | 'single' | 'multiple';
  /**
   * selected map of each item. Default to be selected if item is not in the map
   */
  selected?: Dictionary<boolean>;
  /**
   * Buttons for all select or inverse select.
   * @example
   *  selector: [{type: 'all or inverse', title: xxx}]
   *  selector: true
   *  selector: ['all', 'inverse']
   */
  selector?: (LegendSelectorButtonOption | SelectorType)[] | boolean;
  selectorLabel?: LabelOption;
  emphasis?: {
    selectorLabel?: LabelOption;
  };
  /**
   * Position of selector buttons.
   */
  selectorPosition?: 'auto' | 'start' | 'end';
  /**
   * Gap between each selector button
   */
  selectorItemGap?: number;
  /**
   * Gap between selector buttons group and legend main items.
   */
  selectorButtonGap?: number;
  data?: (string | DataItem)[];
  /**
   * Tooltip option
   */
  tooltip?: CommonTooltipOption<LegendTooltipFormatterParams>;
}
// ...
declare type DimensionUserOuputEncode = {
  [coordOrVisualDimName: string]: DimensionIndex[];
};
// ...
declare type TooltipMarker = string | RichTextTooltipMarker;
// ...
interface TooltipOption extends CommonTooltipOption<TopLevelFormatterParams>, ComponentOption {
  mainType?: 'tooltip';
  axisPointer?: AxisPointerOption & {
    axis?: 'auto' | 'x' | 'y' | 'angle' | 'radius';
    crossStyle?: LineStyleOption & {
      textStyle?: LabelOption;
    };
  };
  /**
   * If show popup content
   */
  showContent?: boolean;
  /**
   * Trigger only works on coordinate system.
   */
  trigger?: 'item' | 'axis' | 'none';
  displayMode?: 'single' | 'multipleByCoordSys';
  /**
   * 'auto': use html by default, and use non-html if `document` is not defined
   * 'html': use html for tooltip
   * 'richText': use canvas, svg, and etc. for tooltip
   */
  renderMode?: 'auto' | TooltipRenderMode;
  /**
   * @deprecated
   * use appendTo: 'body' instead
   */
  appendToBody?: boolean;
  /**
   * If append the tooltip element to another DOM element.
   * Only available when renderMode is html
   */
  appendTo?: ((chartContainer: HTMLElement) => HTMLElement | undefined | null) | string | HTMLElement;
  /**
   * Specify the class name of tooltip element
   * Only available when renderMode is html
   */
  className?: string;
  order?: TooltipOrderMode;
}
// ...
declare type HorizontalAlign = 'left' | 'center' | 'right';
declare type VerticalAlign = 'top' | 'middle' | 'bottom';
declare type ColorString = string;
declare type ZRColor = ColorString | LinearGradientObject | RadialGradientObject | PatternObject;
// ...
declare type DimensionName = string;
// ...
declare type SeriesDataType = 'main' | 'node' | 'edge';
// ...
declare type OptionDataItem =
  | OptionDataValue
  | Dictionary<OptionDataValue>
  | OptionDataValue[]
  | OptionDataItemObject<OptionDataValue>;
declare type OptionDataItemObject<T> = {
  id?: OptionId;
  name?: OptionName;
  groupId?: OptionId;
  childGroupId?: OptionId;
  value?: T[] | T;
  selected?: boolean;
};
declare type OptionId = string | number;
declare type OptionName = string | number;
interface GraphEdgeItemObject<VAL extends OptionDataValue> extends OptionDataItemObject<VAL> {
  /**
   * Name or index of source node.
   */
  source?: string | number;
  /**
   * Name or index of target node.
   */
  target?: string | number;
}
declare type OptionDataValue = string | number | Date | null | undefined;
declare type OptionDataValueNumeric = number | '-';
declare type OptionDataValueDate = Date | string | number;
declare type ModelOption = any;
declare type ThemeOption = Dictionary<any>;
declare type DisplayState = 'normal' | 'emphasis' | 'blur' | 'select';
// ...
interface CallbackDataParams {
  componentType: string;
  componentSubType: string;
  componentIndex: number;
  seriesType?: string;
  seriesIndex?: number;
  seriesId?: string;
  seriesName?: string;
  name: string;
  dataIndex: number;
  data: OptionDataItem;
  dataType?: SeriesDataType;
  value: OptionDataItem | OptionDataValue;
  color?: ZRColor;
  borderColor?: string;
  dimensionNames?: DimensionName[];
  encode?: DimensionUserOuputEncode;
  marker?: TooltipMarker;
  status?: DisplayState;
  dimensionIndex?: number;
  percent?: number;
  $vars: string[];
}
// ...
interface BoxLayoutOptionMixin {
  width?: number | string;
  height?: number | string;
  top?: number | string;
  right?: number | string;
  bottom?: number | string;
  left?: number | string;
}
interface CircleLayoutOptionMixin {
  center?: (number | string)[];
  radius?: (number | string)[] | number | string;
}
interface ShadowOptionMixin {
  shadowBlur?: number;
  shadowColor?: ColorString;
  shadowOffsetX?: number;
  shadowOffsetY?: number;
}
// ...
interface LabelOption extends TextCommonOption {
  /**
   * If show label
   */
  show?: boolean;
  position?: ElementTextConfig['position'];
  distance?: number;
  rotate?: number;
  offset?: number[];
  /**
   * Min margin between labels. Used when label has layout.
   */
  minMargin?: number;
  overflow?: TextStyleProps['overflow'];
  ellipsis?: TextStyleProps['ellipsis'];
  silent?: boolean;
  precision?: number | 'auto';
  valueAnimation?: boolean;
  rich?: Dictionary<TextCommonOption>;
}
// ...
interface TooltipFormatterCallback<T> {
  /**
   * For sync callback
   * params will be an array on axis trigger.
   */
  (params: T, asyncTicket: string): string | HTMLElement | HTMLElement[];
  /**
   * For async callback.
   * Returned html string will be a placeholder when callback is not invoked.
   */
  (
    params: T,
    asyncTicket: string,
    callback: (cbTicket: string, htmlOrDomNodes: string | HTMLElement | HTMLElement[]) => void,
  ): string | HTMLElement | HTMLElement[];
}
declare type TooltipBuiltinPosition = 'inside' | 'top' | 'left' | 'right' | 'bottom';
declare type TooltipBoxLayoutOption = Pick<BoxLayoutOptionMixin, 'top' | 'left' | 'right' | 'bottom'>;
declare type TooltipPositionCallbackParams = CallbackDataParams | CallbackDataParams[];
/**
 * Position relative to the hoverred element. Only available when trigger is item.
 */
interface TooltipPositionCallback {
  (
    point: [number, number],
    /**
     * params will be an array on axis trigger.
     */
    params: TooltipPositionCallbackParams,
    /**
     * Will be HTMLDivElement when renderMode is html
     * Otherwise it's graphic.Text
     */
    el: HTMLDivElement | ZRText | null,
    /**
     * Rect of hover elements. Will be null if not hovered
     */
    rect: RectLike | null,
    size: {
      /**
       * Size of popup content
       */
      contentSize: [number, number];
      /**
       * Size of the chart view
       */
      viewSize: [number, number];
    },
  ): Array<number | string> | TooltipBuiltinPosition | TooltipBoxLayoutOption;
}
/**
 * Common tooltip option
 * Can be configured on series, graphic elements
 */
interface CommonTooltipOption<FormatterParams> {
  show?: boolean;
  /**
   * When to trigger
   */
  triggerOn?: 'mousemove' | 'click' | 'none' | 'mousemove|click';
  /**
   * Whether to not hide popup content automatically
   */
  alwaysShowContent?: boolean;
  formatter?: string | TooltipFormatterCallback<FormatterParams>;
  /**
   * Formatter of value.
   *
   * Will be ignored if tooltip.formatter is specified.
   */
  valueFormatter?: (value: OptionDataValue | OptionDataValue[], dataIndex: number) => string;
  /**
   * Absolution pixel [x, y] array. Or relative percent string [x, y] array.
   * If trigger is 'item'. position can be set to 'inside' / 'top' / 'left' / 'right' / 'bottom',
   * which is relative to the hovered element.
   *
   * Support to be a callback
   */
  position?: (number | string)[] | TooltipBuiltinPosition | TooltipPositionCallback | TooltipBoxLayoutOption;
  confine?: boolean;
  /**
   * Consider triggered from axisPointer handle, verticalAlign should be 'middle'
   */
  align?: HorizontalAlign;
  verticalAlign?: VerticalAlign;
  /**
   * Delay of show. milesecond.
   */
  showDelay?: number;
  /**
   * Delay of hide. milesecond.
   */
  hideDelay?: number;
  transitionDuration?: number;
  /**
   * Whether mouse is allowed to enter the floating layer of tooltip
   * If you need to interact in the tooltip like with links or buttons, it can be set as true.
   */
  enterable?: boolean;
  backgroundColor?: ColorString;
  borderColor?: ColorString;
  borderRadius?: number;
  borderWidth?: number;
  shadowBlur?: number;
  shadowColor?: string;
  shadowOffsetX?: number;
  shadowOffsetY?: number;
  /**
   * Padding between tooltip content and tooltip border.
   */
  padding?: number | number[];
  /**
   * Available when renderMode is 'html'
   */
  extraCssText?: string;
  textStyle?: Pick<
    LabelOption,
    | 'color'
    | 'fontStyle'
    | 'fontWeight'
    | 'fontFamily'
    | 'fontSize'
    | 'lineHeight'
    | 'width'
    | 'height'
    | 'textBorderColor'
    | 'textBorderWidth'
    | 'textShadowColor'
    | 'textShadowBlur'
    | 'textShadowOffsetX'
    | 'textShadowOffsetY'
    | 'align'
  > & {
    decoration?: string;
  };
}
// ...
interface ComponentOption {
  mainType?: string;
  type?: string;
  id?: OptionId;
  name?: OptionName;
  z?: number;
  zlevel?: number;
}
// ...
interface SeriesOption<StateOption = unknown, StatesMixin extends StatesMixinBase = DefaultStatesMixin>
  extends ComponentOption,
    AnimationOptionMixin,
    ColorPaletteOptionMixin,
    StatesOptionMixin<StateOption, StatesMixin> {
  mainType?: 'series';
  silent?: boolean;
  blendMode?: string;
  /**
   * Cursor when mouse on the elements
   */
  cursor?: string;
  /**
   * groupId of data. can be used for doing drilldown / up animation
   * It will be ignored if:
   *  - groupId is specified in each data
   *  - encode.itemGroupId is given.
   */
  dataGroupId?: OptionId;
  data?: unknown;
  colorBy?: ColorBy;
  legendHoverLink?: boolean;
  /**
   * Configurations about progressive rendering
   */
  progressive?: number | false;
  progressiveThreshold?: number;
  progressiveChunkMode?: 'mod';
  /**
   * Not available on every series
   */
  coordinateSystem?: string;
  hoverLayerThreshold?: number;
  /**
   * When dataset is used, seriesLayoutBy specifies whether the column or the row of dataset is mapped to the series
   * namely, the series is "layout" on columns or rows
   * @default 'column'
   */
  seriesLayoutBy?: 'column' | 'row';
  labelLine?: LabelLineOption;
  /**
   * Overall label layout option in label layout stage.
   */
  labelLayout?: LabelLayoutOption | LabelLayoutOptionCallback;
  /**
   * Animation config for state transition.
   */
  stateAnimation?: AnimationOption$1;
  /**
   * If enabled universal transition cross series.
   * @example
   *  universalTransition: true
   *  universalTransition: { enabled: true }
   */
  universalTransition?: boolean | UniversalTransitionOption;
  /**
   * Map of selected data
   * key is name or index of data.
   */
  selectedMap?: Dictionary<boolean> | 'all';
  selectedMode?: 'single' | 'multiple' | 'series' | boolean;
}
// ...
interface EChartsType extends ECharts {}
// ...
interface TitleOption extends ComponentOption, BoxLayoutOptionMixin, BorderOptionMixin {
  mainType?: 'title';
  show?: boolean;
  text?: string;
  /**
   * Link to url
   */
  link?: string;
  target?: 'self' | 'blank';
  subtext?: string;
  sublink?: string;
  subtarget?: 'self' | 'blank';
  textAlign?: ZRTextAlign;
  textVerticalAlign?: ZRTextVerticalAlign;
  /**
   * @deprecated Use textVerticalAlign instead
   */
  textBaseline?: ZRTextVerticalAlign;
  backgroundColor?: ZRColor;
  /**
   * Padding between text and border.
   * Support to be a single number or an array.
   */
  padding?: number | number[];
  /**
   * Gap between text and subtext
   */
  itemGap?: number;
  textStyle?: LabelOption;
  subtextStyle?: LabelOption;
  /**
   * If trigger mouse or touch event
   */
  triggerEvent?: boolean;
  /**
   * Radius of background border.
   */
  borderRadius?: number | number[];
}
// ...
declare type LegendComponentOption = LegendOption | ScrollableLegendOption;